requirements.txt
streamlit>=1.37
scikit-learn>=1.3
pandas>=2.0
numpy>=1.25
scipy>=1.11
PyPDF2>=3.0
python-dotenv>=1.0
requests>=2.31
